CR fixes
