+++ doc/groups.h
@@ -1,11 +1,17 @@
 /**
  * @defgroup perhipheral_drivers Peripheral drivers
+ *
+ * This group contains drivers for MCU peripherals (like I2C, SPI, RTC, DMA)
+ *
  * @{
  * @}
  */
 
 /**
  * @defgroup device_drivers Device drivers
+ *
+ * This group contains drviers for devices connected to OBC (like COMM, EPS, PLD)
+ *
  * @{
  * @}
  */
