+++ libs/base/Include/base/os.h
@@ -199,7 +199,7 @@ enum class TaskPriority
 /**
  * @brief Definition of operating system interface.
  */
-class System : public PureStatic
+class System final : public PureStatic
 {
   public:
     /**
@@ -371,7 +371,7 @@ class System : public PureStatic
      * @param[in] timeout Operation timeout in ms
      * @return TRUE if element was received, FALSE on timeout
      */
-    static bool QueueSend(OSQueueHandle queue, void* element, OSTaskTimeSpan timeout);
+    static bool QueueSend(OSQueueHandle queue, const void* element, OSTaskTimeSpan timeout);
 
     /**
      * @brief Sends element to queue in interrupt handler
@@ -380,7 +380,7 @@ class System : public PureStatic
      * @param[in] element Element to send to queue
      * @return TRUE if element was received, FALSE on timeout
      */
-    static bool QueueSendISR(OSQueueHandle queue, void* element);
+    static bool QueueSendISR(OSQueueHandle queue, const void* element);
 
     /**
      * @brief Overwrites element in queue
@@ -422,7 +422,7 @@ class System : public PureStatic
 /**
  * RTOS Task wrapper
  */
-template <typename Param, std::uint16_t StackSize, TaskPriority Priority> class Task
+template <typename Param, std::uint16_t StackSize, TaskPriority Priority> class Task final
 {
     static_assert(sizeof(Param) < 16, "WTF are you trying to do?");
     static_assert(StackSize % 2 == 0, "Stack size must even");
@@ -490,7 +490,7 @@ template <typename Param, std::uint16_t StackSize, TaskPriority Priority> void T
  *
  * @code
  * {
- * 	TakeSemaphore lock(this->_sem);
+ * 	Lock lock(this->_sem);
  *
  * 	if(!lock())
  * 	{
@@ -502,20 +502,20 @@ template <typename Param, std::uint16_t StackSize, TaskPriority Priority> void T
  * } // semaphore release at the end of scope
  * @endcode
  */
-class TakeSemaphore
+class Lock final
 {
   public:
     /**
-     * @brief Constructs @ref TakeSemaphore object and tries to acquire semaphore
+     * @brief Constructs @ref Lock object and tries to acquire semaphore
      * @param[in] semaphore Semaphore to take
      * @param[in] timeout Timeout
      */
-    TakeSemaphore(OSSemaphoreHandle semaphore, OSTaskTimeSpan timeout);
+    Lock(OSSemaphoreHandle semaphore, OSTaskTimeSpan timeout);
 
     /**
      * @brief Releases semaphore (if taken) on object destruction
      */
-    ~TakeSemaphore();
+    ~Lock();
 
     /**
      * @brief Checks if semaphore has been taken
@@ -523,9 +523,12 @@ class TakeSemaphore
      */
     bool operator()();
 
-    TakeSemaphore(const TakeSemaphore&) = delete;
-
   private:
+    Lock(const Lock&) = delete;
+    Lock& operator=(const Lock&) = delete;
+    Lock(Lock&&) = delete;
+    Lock& operator=(Lock&&) = delete;
+
     /** @brief Semaphore handle */
     const OSSemaphoreHandle _semaphore;
     /** @brief Flag indicating if semaphore is acquired */
@@ -535,8 +538,10 @@ class TakeSemaphore
 /**
  * @brief RTOS queue wrapper
  */
-template <typename Element, std::size_t Capacity> class Queue
+template <typename Element, std::size_t Capacity> class Queue final
 {
+    static_assert(std::is_pod<Element>::value, "Queue works only for POD/integral types");
+
   public:
     /**
      * @brief Creates underlying RTOS queue object
@@ -550,14 +555,14 @@ template <typename Element, std::size_t Capacity> class Queue
      * @param[in] timeout Timeout in ms
      * @return Operation result
      */
-    OSResult Push(Element* element, OSTaskTimeSpan timeout);
+    OSResult Push(const Element& element, OSTaskTimeSpan timeout);
 
     /**
      * @brief Pushes element to queue from interrupt service routine
      * @param[in] element Pointer to element that will be placed in queue
      * @return Operation result
      */
-    OSResult PushISR(Element* element);
+    OSResult PushISR(const Element& element);
 
     /**
      * @brief Pops element from queue
@@ -565,7 +570,7 @@ template <typename Element, std::size_t Capacity> class Queue
      * @param[in] timeout Timeout in ms
      * @return Operation result
      */
-    OSResult Pop(Element* element, OSTaskTimeSpan timeout);
+    OSResult Pop(Element& element, OSTaskTimeSpan timeout);
 
   private:
     /** @brief Queue handle */
@@ -584,27 +589,27 @@ template <typename Element, std::size_t Capacity> OSResult Queue<Element, Capaci
     return OSResult::Success;
 }
 
-template <typename Element, std::size_t Capacity> OSResult Queue<Element, Capacity>::Push(Element* element, OSTaskTimeSpan timeout)
+template <typename Element, std::size_t Capacity> OSResult Queue<Element, Capacity>::Push(const Element& element, OSTaskTimeSpan timeout)
 {
-    if (System::QueueSend(this->_handle, element, timeout))
+    if (System::QueueSend(this->_handle, &element, timeout))
     {
         return OSResult::Success;
     }
     return OSResult::Timeout;
 }
 
-template <typename Element, std::size_t Capacity> OSResult Queue<Element, Capacity>::PushISR(Element* element)
+template <typename Element, std::size_t Capacity> OSResult Queue<Element, Capacity>::PushISR(const Element& element)
 {
-    if (System::QueueSendISR(this->_handle, element))
+    if (System::QueueSendISR(this->_handle, &element))
     {
         return OSResult::Success;
     }
     return OSResult::Overflow;
 }
 
-template <typename Element, std::size_t Capacity> OSResult Queue<Element, Capacity>::Pop(Element* element, OSTaskTimeSpan timeout)
+template <typename Element, std::size_t Capacity> OSResult Queue<Element, Capacity>::Pop(Element& element, OSTaskTimeSpan timeout)
 {
-    if (System::QueueReceive(this->_handle, element, timeout))
+    if (System::QueueReceive(this->_handle, &element, timeout))
     {
         return OSResult::Success;
     }
