+++ libs/base/os_base.cpp
@@ -1,18 +1,11 @@
 #include "os.h"
 
-TakeSemaphore::TakeSemaphore(OSSemaphoreHandle semaphore, OSTaskTimeSpan timeout) : _semaphore(semaphore)
+Lock::Lock(OSSemaphoreHandle semaphore, OSTaskTimeSpan timeout) : _semaphore(semaphore)
 {
-    if (OS_RESULT_SUCCEEDED(System::TakeSemaphore(this->_semaphore, timeout)))
-    {
-        this->_taken = true;
-    }
-    else
-    {
-        this->_taken = false;
-    }
+    this->_taken = OS_RESULT_SUCCEEDED(System::TakeSemaphore(this->_semaphore, timeout));
 }
 
-TakeSemaphore::~TakeSemaphore()
+Lock::~Lock()
 {
     if (this->_taken)
     {
@@ -20,7 +13,7 @@ TakeSemaphore::~TakeSemaphore()
     }
 }
 
-bool TakeSemaphore::operator()()
+bool Lock::operator()()
 {
     return this->_taken;
 }
