+++ libs/drivers/i2c/Include/i2c/i2c.h
@@ -97,87 +97,6 @@ namespace drivers
             virtual I2CResult WriteRead(const I2CAddress address, gsl::span<const uint8_t> inData, gsl::span<uint8_t> outData) = 0;
         };
 
-        /** @brief Low-level I2C bus driver */
-        class I2CLowLevelBus : public II2CBus
-        {
-          public:
-            /**
-             * @brief Setups I2C low-level driver object. No RTOS or hardware initialization is done in constructor
-             * @param[in] hw I2C hardware registers set
-             * @param[in] location Pins location to use
-             * @param[in] port GPIO port to use
-             * @param[in] sdaPin Number of GPIO pin to use for SDA line
-             * @param[in] sclPin Number of GPIO pin to use for SCL line
-             * @param[in] clock Clock used by selected hardware interface
-             * @param[in] irq IRQ number used by selected hardware interface
-             */
-            I2CLowLevelBus(I2C_TypeDef* hw,
-                uint16_t location,
-                GPIO_Port_TypeDef port,
-                uint16_t sdaPin,
-                uint16_t sclPin,
-                CMU_Clock_TypeDef clock,
-                IRQn_Type irq);
-
-            virtual I2CResult Write(const I2CAddress address, gsl::span<const uint8_t> inData) override;
-
-            virtual I2CResult WriteRead(const I2CAddress address, gsl::span<const uint8_t> inData, gsl::span<uint8_t> outData) override;
-
-            /**
-             * @brief Initializes RTOS and hardware components of I2C low-level driver
-             */
-            void Initialize();
-
-            /**
-             * @brief Interrupt handler for I2C hardware
-             */
-            void IRQHandler();
-
-          private:
-            /**
-             * @brief Executes single I2C transfer
-             * @param[in] bus I2C bus
-             * @param[in] seq Transfer sequence definition
-             * @return Transfer result
-             */
-            I2CResult ExecuteTransfer(I2C_TransferSeq_TypeDef* seq);
-
-            /**
-             * @brief Checks if SCL line is latched at low level
-             * @param[in] bus I2C bus
-             * @return true if SCL line is latched
-             */
-            bool IsSclLatched();
-
-            /** @brief Pointer to hardware registers */
-            void* HWInterface;
-
-            /**
-             * @brief GPIO used by this I2C interface
-             */
-            struct
-            {
-                /** @brief Peripheral clock */
-                CMU_Clock_TypeDef Clock;
-                /** @brief Peripheral interrupt number */
-                IRQn_Type IRQn;
-                /** @brief Peripheral location number */
-                uint16_t Location;
-                /** @brief Used port */
-                GPIO_Port_TypeDef Port;
-                /** @brief Number of pin connected to SCL line */
-                uint16_t SCL;
-                /** @brief Number of pin connected to SDA line */
-                uint16_t SDA;
-            } _io;
-
-            /** @brief Lock used to synchronize access to bus */
-            OSSemaphoreHandle _lock;
-
-            /** @brief Single-element queue storing results of transfers */
-            Queue<I2C_TransferReturn_TypeDef, 1> _resultQueue;
-        };
-
         /**
          * @brief Type with references to used I2C buses
          */
@@ -193,66 +112,6 @@ namespace drivers
             II2CBus& Bus;
             /** @brief Reference to Payload I2C bus */
             II2CBus& Payload;
-        };
-
-        /**
-         * @brief I2C Fallbacking bus driver
-         * @implements I2CBus
-         */
-        class I2CFallbackBus final : public II2CBus
-        {
-          public:
-            /**
-             * @brief Setups bus wrapper that fallbacks from system to payload bus in case of failure
-             * @param[in] buses Object representing both buses used in the system
-             */
-            I2CFallbackBus(I2CInterface& buses);
-
-            virtual I2CResult Write(const I2CAddress address, gsl::span<const uint8_t> inData) override;
-
-            virtual I2CResult WriteRead(const I2CAddress address, gsl::span<const uint8_t> inData, gsl::span<uint8_t> outData) override;
-
-          private:
-            /** @brief Object representing both buses used in the system */
-            I2CInterface& _innerBuses;
-        };
-
-        /**
-         * @brief Type of procedure used by error handling bus wrapper
-         * @param[in] bus Bus on which transfer failed
-         * @param[in] result Transfer error code
-         * @param[in] address Device that was addressed
-         * @param[in] context Context
-         * @return New result code
-         */
-        using BusErrorHandler = I2CResult (*)(II2CBus& bus, I2CResult result, I2CAddress address, void* context);
-
-        /**
-         * @brief Error handling bus driver
-         * @implements I2CBus
-         */
-        class I2CErrorHandlingBus final : public II2CBus
-        {
-          public:
-            /**
-             * @brief Initializes new instance of @ref I2CErrorHandlingBus
-             * @param[in] innerBus Bus that will be wrapped
-             * @param[in] handler Pointer to error handler function
-             * @param[in] context Parameter passed to error handler function
-             */
-            I2CErrorHandlingBus(II2CBus& innerBus, BusErrorHandler handler, void* context);
-
-            virtual I2CResult Write(const I2CAddress address, gsl::span<const uint8_t> inData) override;
-
-            virtual I2CResult WriteRead(const I2CAddress address, gsl::span<const uint8_t> inData, gsl::span<uint8_t> outData) override;
-
-          private:
-            /** @brief Underlying bus */
-            II2CBus& _innerBus;
-            /** @brief Pointer to function called in case of error */
-            const BusErrorHandler _errorHandler;
-            /** @brief Context passed to error handler function */
-            void* _handlerContext;
         };
 
         /** @} */
