+++ unit_tests/CMakeLists.txt
@@ -24,7 +24,6 @@ set(SOURCES
   MissionPlan/antenna/AntennaMissionStateTest.cpp
   I2C/FallbackI2CBusTest.cpp
   I2C/ErrorHandlingI2CBusTest.cpp
-  I2C/I2CMock.cpp
   os/os.cpp
   os/os.hpp
   OsMock.hpp
