--- unit_tests/I2C/I2CMock.cpp
@@ -1,5 +1,0 @@
-#include "I2CMock.hpp"
-
-I2CBusMock::I2CBusMock()
-{
-}
